Update interpose to fix ShapeLayerHelper update issue
